Changed the German translations. DeepL had produced some *very* funny-sounding translations, like translating "plant" (as in, trees and flowers) to the German word "Anlage", which means "facility" or "sound system". (DeepL's thought process being: "plant" means "factory" or "facility", "facility" means "Anlage".) There were a lot of these interesting tidbits.
Also, I am still convinced no one will ever know what "Filmrolle" is
supposed to be without trying it out. It is the direct translation of
"Camera Roll", but no one in Germany uses this term to refer to
something digital. Sadly, even as a native speaker I cannot think of a
better term.